[v0.90.1][tests] Split provider_tests.rs responsibilities (Closes #2209)
